project6-safe-rl/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
Farama-Notifications==0.0.4
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
glfw==2.10.0
grpcio==1.78.0
gymnasium==0.28.1
idna==3.11
ImageIO==2.37.3
imageio-ffmpeg==0.6.0
jax-jumpy==1.0.0
Jinja2==3.1.6
kiwisolver==1.5.0
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
mujoco==2.3.0
networkx==3.4.2
numpy==1.26.4
omnisafe==0.5.0
packaging==26.0
pandas==2.0.3
pillow==11.3.0
platformdirs==4.9.4
proglog==0.1.12
protobuf==6.33.6
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.1.0
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
safety-gymnasium==0.4.1
seaborn==0.13.2
sentry-sdk==2.55.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.10.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.25.1
Werkzeug==3.1.6
xmltodict==1.0.4
